reintegrated disabled tests
