+++ planetmint/utils.py
@@ -10,6 +10,15 @@ import multiprocessing
 import setproctitle
 
 
+class Singleton(type):
+    _instances = {}
+
+    def __call__(cls, *args, **kwargs):
+        if cls not in cls._instances:
+            cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
+        return cls._instances[cls]
+
+
 class Process(multiprocessing.Process):
     """Wrapper around multiprocessing.Process that uses
     setproctitle to set the name of the process when running
